python_scripts/raydium_py/raydium_py/requirements.txt
python-dotenv>=0.21.0
solana==0.35.0
solders==0.21.0
